% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wagi.R
\name{oblicz_wagi_entropia}
\alias{oblicz_wagi_entropia}
\title{Wagi obiektywne metoda entropii Shannona}
\usage{
oblicz_wagi_entropia(macierz)
}
\arguments{
\item{macierz}{Macierz liczbowa o nieujemnych wartosciach: wiersze to
alternatywy, kolumny to kryteria.}
}
\value{
Nazwany wektor wag sumujacych sie do jednosci.
}
\description{
Kryterium, w ktorym oceny alternatyw sa do siebie zblizone, niesie malo
informacji rozniczujacej i otrzymuje wage nizsza. Procedura sklada sie
z trzech krokow: normalizacji sumacyjnej, wyznaczenia entropii kolumny
oraz zamiany rozbieznosci na wagi sumujace sie do jednosci.
}
\details{
Dla macierzy o \eqn{m} alternatywach entropia kolumny \eqn{j} wynosi
\eqn{e_j = -k \sum_i p_{ij} \ln p_{ij}}, gdzie \eqn{k = 1 / \ln(m)},
a \eqn{p_{ij}} to udzial alternatywy w sumie kolumny. Waga powstaje
z rozbieznosci \eqn{d_j = 1 - e_j} po znormalizowaniu do sumy jeden.

Kolumna o zerowym rozrzucie ma entropie rowna jednosci, wiec jej
rozbieznosc wynosi zero. Gdy wszystkie kolumny sa stale, wagi nie daja
sie wyznaczyc i funkcja zwraca rozklad rownomierny.
}
\examples{
m <- matrix(c(1, 2, 3, 4, 4, 4), nrow = 3,
            dimnames = list(NULL, c("a", "b")))
oblicz_wagi_entropia(m)

}
