% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wagi.R
\name{oblicz_wagi_std}
\alias{oblicz_wagi_std}
\title{Wagi obiektywne oparte na odchyleniu standardowym}
\usage{
oblicz_wagi_std(macierz)
}
\arguments{
\item{macierz}{Macierz liczbowa o nieujemnych wartosciach: wiersze to
alternatywy, kolumny to kryteria.}
}
\value{
Nazwany wektor wag sumujacych sie do jednosci.
}
\description{
Waga kryterium jest proporcjonalna do odchylenia standardowego jego ocen.
Metoda jest prostsza od entropii i sluzy jako punkt odniesienia przy
sprawdzaniu, czy wynik rankingu nie zalezy nadmiernie od sposobu wazenia.
}
\examples{
m <- matrix(c(1, 2, 3, 4, 4, 4), nrow = 3,
            dimnames = list(NULL, c("a", "b")))
oblicz_wagi_std(m)

}
