% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizualizacja.R
\name{plot.ranking_epi}
\alias{plot.ranking_epi}
\title{Wykres rankingu alternatyw}
\usage{
\method{plot}{ranking_epi}(x, tytul = "Ranking alternatyw", ...)
}
\arguments{
\item{x}{Obiekt klasy \code{ranking_epi}.}

\item{tytul}{Tytul wykresu.}

\item{...}{Argumenty przekazywane dalej, obecnie nieuzywane.}
}
\value{
Obiekt klasy \code{ggplot}.
}
\description{
Wykres slupkowy uporzadkowany wedlug wskaznika. Alternatywy o wskazniku
zblizonym do siebie widac na nim od razu, co jest istotniejsze niz sama
kolejnosc: bliskie slupki oznaczaja, ze roznica moze wynikac z szumu
w danych, a nie z rzeczywistej przewagi.
}
\examples{
macierz <- przygotuj_dane(dane_przykladowe, "Koszt =~ koszt_surowce")
plot(oblicz_ranking(macierz))

}
