% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klasy_s3.R
\name{print.macierz_decyzyjna}
\alias{print.macierz_decyzyjna}
\title{Wypisanie macierzy decyzyjnej}
\usage{
\method{print}{macierz_decyzyjna}(x, ...)
}
\arguments{
\item{x}{Obiekt klasy \code{macierz_decyzyjna}.}

\item{...}{Argumenty przekazywane dalej, obecnie nieuzywane.}
}
\value{
Niewidocznie zwraca \code{x}. Wywolywana dla efektu ubocznego.
}
\description{
Wypisanie macierzy decyzyjnej
}
\examples{
macierz <- przygotuj_dane(dane_przykladowe, "Koszt =~ koszt_surowce")
print(macierz)

}
