% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klasy_s3.R
\name{print.ranking_epi}
\alias{print.ranking_epi}
\title{Wypisanie rankingu}
\usage{
\method{print}{ranking_epi}(x, ...)
}
\arguments{
\item{x}{Obiekt klasy \code{ranking_epi}.}

\item{...}{Argumenty przekazywane dalej, obecnie nieuzywane.}
}
\value{
Niewidocznie zwraca \code{x}. Wywolywana dla efektu ubocznego.
}
\description{
Wypisanie rankingu
}
\examples{
macierz <- przygotuj_dane(dane_przykladowe, "Koszt =~ koszt_surowce")
print(oblicz_ranking(macierz))

}
