% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{przygotuj_dane}
\alias{przygotuj_dane}
\title{Przygotowanie macierzy decyzyjnej z surowej ramki danych}
\usage{
przygotuj_dane(
  dane,
  model,
  alternatywa = "alternatywa",
  braki = c("mediana", "usun", "blad"),
  kody_bledne = c(-999, 999)
)
}
\arguments{
\item{dane}{Ramka danych z kolumna identyfikujaca alternatywy oraz
zmiennymi surowymi.}

\item{model}{Znakowy opis modelu w skladni opisanej wyzej.}

\item{alternatywa}{Nazwa kolumny identyfikujacej alternatywy.}

\item{braki}{Sposob obslugi brakow danych: \code{"mediana"} zastepuje brak
mediana kolumny, \code{"usun"} usuwa wiersz, \code{"blad"} zatrzymuje
wykonanie.}

\item{kody_bledne}{Wektor wartosci traktowanych jako kod bledu i zamienianych
na brak danych przed dalszym przetwarzaniem.}
}
\value{
Obiekt klasy \code{macierz_decyzyjna}: macierz o wierszach
  odpowiadajacych alternatywom i kolumnach odpowiadajacych kryteriom,
  z atrybutami \code{skladniki} oraz \code{liczba_obserwacji}.
}
\description{
Funkcja przeprowadza surowe dane przez cztery kroki: sprawdzenie warunkow
wstepnych, zamiane kodow bledych na braki danych, obsluge brakow oraz
agregacje zmiennych do kryteriow zadeklarowanych w modelu.
}
\details{
Model zapisuje sie w skladni zblizonej do pakietu lavaan. Kazdy wiersz
deklaruje jedno kryterium i wymienia zmienne, ktore sie na nie skladaja:

\preformatted{
Koszt  =~ koszt_surowce + koszt_praca
Jakosc =~ jakosc_trwalosc + jakosc_ux
}
}
\examples{
model <- "Koszt  =~ koszt_surowce + koszt_praca
          Jakosc =~ jakosc_trwalosc + jakosc_ux"
macierz <- przygotuj_dane(dane_przykladowe, model)
macierz

}
