% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{rozbierz_model}
\alias{rozbierz_model}
\title{Rozbior opisu modelu na kryteria i zmienne}
\usage{
rozbierz_model(model)
}
\arguments{
\item{model}{Znakowy opis modelu w skladni opisanej w \code{\link{przygotuj_dane}}.}
}
\value{
Nazwana lista wektorow znakowych. Nazwy sa kryteriami, elementy
  nazwami zmiennych skladowych.
}
\description{
Rozbior opisu modelu na kryteria i zmienne
}
\examples{
rozbierz_model("Koszt =~ a + b")

}
