% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{skaluj_min_max}
\alias{skaluj_min_max}
\title{Skalowanie wektora do przedzialu jednostkowego}
\usage{
skaluj_min_max(x)
}
\arguments{
\item{x}{Wektor liczbowy.}
}
\value{
Wektor liczbowy o wartosciach z przedzialu od 0 do 1.
}
\description{
Zmienna o zerowym rozrzucie otrzymuje wartosc stala rowna 0,5, poniewaz
nie rozroznia alternatyw, a jej usuniecie zmienialoby model zadeklarowany
przez uzytkownika.
}
\examples{
skaluj_min_max(c(2, 4, 6))

}
