% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klasy_s3.R
\name{summary.ranking_epi}
\alias{summary.ranking_epi}
\title{Podsumowanie rankingu}
\usage{
\method{summary}{ranking_epi}(object, ...)
}
\arguments{
\item{object}{Obiekt klasy \code{ranking_epi}.}

\item{...}{Argumenty przekazywane dalej, obecnie nieuzywane.}
}
\value{
Niewidocznie zwraca liste ze skladnikami \code{wynik}, \code{wagi}
  i \code{rozpietosc}. Wywolywana glownie dla efektu ubocznego.
}
\description{
Poza uporzadkowana lista alternatyw podsumowanie podaje wagi kryteriow
oraz rozpietosc wskaznika, ktora informuje, jak wyraznie ranking rozroznia
alternatywy. Rozpietosc bliska zeru oznacza, ze wynik jest wrazliwy na
drobne zmiany danych.
}
\examples{
macierz <- przygotuj_dane(dane_przykladowe, "Koszt =~ koszt_surowce")
summary(oblicz_ranking(macierz))

}
