% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SzablonR-package.R
\docType{package}
\name{SzablonR-package}
\alias{SzablonR}
\alias{SzablonR-package}
\title{SzablonR: szkielet pakietu badawczego}
\description{
Pakiet stanowi punkt wyjscia do wlasnej implementacji algorytmu opisanego
w artykule metodycznym. Pokazuje uklad, ktorego oczekuje sprawdzanie
zgodnosci z repozytorium: rozdzielenie przygotowania danych od silnika
obliczeniowego, jawny kontrakt bledow, klase wyniku z metodami generycznymi
oraz procedure generowania danych o znanej strukturze.
}
\details{
Sciezka analizy przebiega przez trzy funkcje. \code{\link{przygotuj_dane}}
zamienia surowa ramke danych w macierz decyzyjna. \code{\link{oblicz_ranking}}
wyznacza wagi i wskaznik. Metody \code{print}, \code{summary} i \code{plot}
pokazuja wynik.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nazwauzytkownika/SzablonR}
  \item Report bugs at \url{https://github.com/nazwauzytkownika/SzablonR/issues}
}

}
\author{
\strong{Maintainer}: Imie Nazwisko \email{imie.nazwisko@student.uj.edu.pl}

}
\keyword{internal}
