% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizualizacja.R
\name{wykres_udzialow}
\alias{wykres_udzialow}
\title{Wykres udzialu kryteriow we wskazniku}
\usage{
wykres_udzialow(x, tytul = "Udzial kryteriow we wskazniku")
}
\arguments{
\item{x}{Obiekt klasy \code{ranking_epi}.}

\item{tytul}{Tytul wykresu.}
}
\value{
Obiekt klasy \code{ggplot}.
}
\description{
Rozklada wskaznik kazdej alternatywy na wklady poszczegolnych kryteriow.
Pozwala odpowiedziec na pytanie, dlaczego dana alternatywa zajela swoja
pozycje, a nie tylko jaka pozycje zajela.
}
\examples{
model <- "Koszt  =~ koszt_surowce + koszt_praca
          Jakosc =~ jakosc_trwalosc + jakosc_ux"
macierz <- przygotuj_dane(dane_przykladowe, model)
wykres_udzialow(oblicz_ranking(macierz))

}
